Heavily restructure ::wp::option and ::wp::plugin (v1.0.2)
This commit makes it possible to run "wp" as root user, which is necessary when
working with operating system -provided Wordpress packages. Dependencies to
wp::command are removed from option.pp and plugin.pp, tabs are replaced with
spaces and some parameters are moved to the wp::params class to reduce
repetition. In addition defining $location globally is now possible using the
wp::location parameter.

Signed-off-by: Samuli Seppänen <samuli.seppanen@gmail.com>